requirements.txt
djangorestframework-jwt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-heroku==0.3.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
whitenoise==5.0.1
